Core/PacketIO: Use new time utilities in packets instead of plain time_t/uint32

+++ src/server/game/Server/Packets/MiscPackets.cpp
@@ -756,8 +756,8 @@ void WorldPackets::Misc::AdventureJournalStartQuest::Read()
 
 WorldPacket const* WorldPackets::Misc::StartTimer::Write()
 {
-    _worldPacket << int32(TimeLeft);
-    _worldPacket << int32(TotalTime);
+    _worldPacket << TimeLeft;
+    _worldPacket << TotalTime;
     _worldPacket << int32(Type);
 
     return &_worldPacket;
